refactor(core): extract pipeline geometry, facets, and shared types
Split the 3.7k-line pipeline monolith into focused modules under
pipeline/{types,geometry,facets}.ts while preserving the public
pipeline.ts import surface. Characterization tests pin batch mark
counts, coord-flip vertex mapping, and facet partition contracts.
